+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.10.0
 
-- Add ramda 0.27.2 to resolutions [#256](https://github.com/grafana/grafana-iot-twinmaker-app/pull/256)
+- Config editor: Fix workspace loading in new form styling in [#258](https://github.com/grafana/grafana-iot-twinmaker-app/pull/258)
+- Add ramda 0.27.2 to resolutions in [#256](https://github.com/grafana/grafana-iot-twinmaker-app/pull/256)
 - Query editor: Migrate to new form styling under feature toggle in [#249](https://github.com/grafana/grafana-iot-twinmaker-app/pull/249)
 - Config editor: Migrate to new form styling under feature toggle in [#244](https://github.com/grafana/grafana-iot-twinmaker-app/pull/244)